requirements.txt
absl-py==0.13.0
aniso8601==9.0.1
anyio==3.3.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.0.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==7.1.1
colorama==0.4.4
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
efficientnet-pytorch==0.6.3
entrypoints==0.3
enum34==1.1.10
Flask==1.0
Flask-RESTful==0.3.9
flatbuffers==1.12
future==0.18.2
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.1.0
idna==2.10
immutables==0.16
importlib-metadata==4.6.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mock==4.0.3
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.18.4
oauthlib==3.1.1
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==21.0
pandas==1.0.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
piexif==1.1.3
Pillow==8.1.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.2
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2020.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4
pyzmq==22.2.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
Send2Trash==1.8.0
six==1.14.0
sniffio==1.2.0
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==0.15.5
wrapt==1.12.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.5.0
